Notification is working, sort of.
It (sometimes?) triggers an infinite loop of settings updates, and if the extension weren’t running at the time, we’d not get any notification.  Essentially, the current mechanism should be ripped out.

On-demand launch of our extension seems to be a feature, not a bug.  However, it seems wasteful to require launching/XPC on potentially EVERY page load just to figure out whether we should enable ourselves on that page.

Defining a separate “prepare” event which loads settings might be the best idea.